+++ crates/miden-protocol/src/block/mod.rs
@@ -7,9 +7,6 @@ pub use block_body::BlockBody;
 mod block_number;
 pub use block_number::BlockNumber;
 
-mod block_proof;
-pub use block_proof::BlockProof;
-
 mod kernel;
 pub use kernel::BlockKernel;